[sdk]: move phantom aggregation into the SDK; assert it in the simplex E2E
Moves aggregatePhantomBids + the phantom simulation helpers from the indexer into
@hyperbridge/sdk (exported from the VM2-safe intents-helpers entry), with the two
indexer-specific config maps — per-token ERC-20 slots and the ERC-4626 vault map —
passed in as parameters. This lets both the indexer handler and the simplex E2E
share one implementation (the earlier cross-package @/-alias collision is gone).

- sdk: new src/protocols/intents/phantom-aggregation.ts; re-exported from
  intents-helpers (bytes32->address inlined to keep that entry light).
- indexer: handler imports aggregatePhantomBids from the SDK and passes
  TOKEN_SLOT_OVERRIDES + YIELD_VAULT_ADDRESSES; deletes the local copies; the helper
  unit test imports from the SDK and threads the overrides map.
- simplex: the phantom-filler E2E now also runs aggregatePhantomBids on the real
  bids and asserts the snapshot (bidCount, price range, weighted median, cNGN LP
  balances) — full flow in one test.

Validated locally: SDK build (incl. dts type-check) passes; runtime check of the
moved functions from the built bundle passes. The simnode+anvil E2E still needs
those services to run.
