+++ sdk/packages/sdk/src/intents-helpers.ts
@@ -3,3 +3,27 @@
 export { decodeERC7821ExecuteBatch, encodeERC7821ExecuteBatch } from "@/protocols/intents/decode-utils"
 export { decodeUserOpScale, encodeUserOpScale } from "@/chains/intentsCoprocessor"
 export { default as IntentGatewayV2 } from "@/abis/IntentGatewayV2"
+export {
+	aggregatePhantomBids,
+	buildSimulationOrder,
+	erc20AllowanceSlot,
+	erc20BalanceSlot,
+	extractFillData,
+	fetchBidsForOrder,
+	getTotalSolverBalance,
+	hasTokenSlotOverride,
+	ordersStorageSlot,
+	tokenSlots,
+	weightedMedian,
+	FILL_ORDER_ABI,
+	ORDER_FILLED_TOPIC,
+	SIM_DEADLINE,
+	type AggregationLogger,
+	type FillData,
+	type LpBalance,
+	type PhantomAggregation,
+	type RpcBidInfo,
+	type TokenSlots,
+	type TokenSlotOverrides,
+	type YieldVaultMap,
+} from "@/protocols/intents/phantom-aggregation"
